Moved some components into an archive folder.
