+++ test/AdjacentPalette.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import AdjacentPalette from '../src/components/AdjacentPalette';
+import AdjacentPalette from '../src/components/old-components/AdjacentPalette';
 
 describe('AdjacentPalette', () => {
     it('renders the correct title and contrast ratio', () => {
